requirements.txt
pandas
pyyaml
aqxle 